define and enumerate fungi, and establish mud mixing mechanics
traits are defined in frog.py, but all modifications to traits are defined
here as they are controlled by fungi

TODO: still have not made the leap from a mushroom modifier string to a
single or multiple trait modifier object.

single trait modifiers are e.g. "increase amplitude by 1", and multiple trait
modifiers encompass multiple things one mushroom or a mud mixture might include.

modifiers are aggregated in multi, but not flattened until mud mixing, as part
of the mud class.

a mushroom is defined here. parsing its effects from the string is incomplete.

all mushrooms in the game are enumerated here.